refactor: usePatchPickData 리팩토링
