+++ pages/pickpickpick/modify/[id]/apiHooks/usePatchPickData.ts
@@ -1,7 +1,12 @@
 import axios from 'axios';
 
-import { useMutation } from '@tanstack/react-query';
+import { useRouter } from 'next/router';
 
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+
+import { PICK_UPDATE_MESSAGE } from '@pages/pickposting/constants/pickPostConstants';
+
+import { useModalStore } from '@stores/modalStore';
 import { useToastVisibleStore } from '@stores/toastVisibleStore';
 
 import { UNDEFINED_ERROR_MESSAGE } from '@/constants/errorMessageConstants';
@@ -20,11 +25,28 @@ export const patchPickData = async ({
 };
 
 export const usePatchPickData = () => {
+  const router = useRouter();
+  const { id } = router.query;
+
   const { setToastVisible } = useToastVisibleStore();
+  const { popModal } = useModalStore();
+
+  const queryClient = useQueryClient();
 
   return useMutation({
     mutationFn: patchPickData,
+    onSuccess: () => {
+      popModal();
+      queryClient.invalidateQueries({
+        queryKey: ['getDetailPickData', id],
+      });
+      queryClient.invalidateQueries({ queryKey: ['pickData'] });
+      queryClient.invalidateQueries({ queryKey: ['myPicksData'] });
+      router.push(`/pickpickpick/${id}`);
+      setToastVisible({ message: PICK_UPDATE_MESSAGE });
+    },
     onError: (error: ErrorRespone) => {
+      popModal();
       const errorMessage = error.response.data.message;
 
       if (errorMessage == null) {
