+++ pages/pickpickpick/modify/[id]/index.page.tsx
@@ -1,12 +1,6 @@
 import { useRouter } from 'next/router';
 
-import { useQueryClient } from '@tanstack/react-query';
-
 import { useGetPickDetailData } from '@pages/pickpickpick/[id]/apiHooks/usePickDetailData';
-import { PICK_UPDATE_MESSAGE } from '@pages/pickposting/constants/pickPostConstants';
-
-import { useModalStore } from '@stores/modalStore';
-import { useToastVisibleStore } from '@stores/toastVisibleStore';
 
 import PickForm from '@components/features/pickpickpick/PickForm';
 import { MutatePickProps } from '@components/features/pickpickpick/types/formPicks';
@@ -18,33 +12,14 @@ export default function Index() {
 
   const { id } = router.query;
 
-  const { popModal } = useModalStore();
-
   const { data: pickDetailData } = useGetPickDetailData(id as string);
-  const { mutate: patchPickMutate, isPending } = usePatchPickData();
-  const { setToastVisible } = useToastVisibleStore();
-  const queryClient = useQueryClient();
+  const { mutate: patchPickMutate } = usePatchPickData();
 
   const handleUpdateSubmit = (pickData: MutatePickProps) => {
-    popModal(); // FIXME: 에러일 때 토스트 띄우기로 변경
-    patchPickMutate(
-      {
-        id,
-        pickData,
-      },
-      {
-        onSuccess: () => {
-          popModal();
-          queryClient.invalidateQueries({
-            queryKey: ['getDetailPickData', id],
-          });
-          queryClient.invalidateQueries({ queryKey: ['pickData'] });
-          queryClient.invalidateQueries({ queryKey: ['myPicksData'] });
-          router.push(`/pickpickpick/${id}`);
-          setToastVisible({ message: PICK_UPDATE_MESSAGE });
-        },
-      },
-    );
+    patchPickMutate({
+      id,
+      pickData,
+    });
   };
 
   return (
